move busan_beach to water quality
