+++ ai4water/datasets/water_quality/__init__.py
@@ -1,4 +1,5 @@
 
 from ._doce_river import DoceRiver
+from ._busan_beach import busan_beach
 from ._selune_river import SeluneRiver
 from ._river_chem_siberia import RiverChemSiberia
